Implement forum mocking and adapt existing tests
