+++ app/src/main/java/com/github/h3lp3rs/h3lp/database/FireDatabase.kt
@@ -10,7 +10,7 @@ import java.util.concurrent.CompletableFuture
 /**
  * Implementation of a NoSQL external database based on Firebase
  */
-internal class FireDatabase(path : String) : Database {
+class FireDatabase(path : String) : Database {
 
     private val db : DatabaseReference =
         Firebase.database("https://h3lp-signin-default-rtdb.europe-west1.firebasedatabase.app/").reference.child(
